Updating to patch for HBASE-7958
That patch is a little more advanced in terms of flexibility in readers/writers and what the statistictracker
actually needs to do. Here we co-opt most of that behavior and munge it a little bit to support the
coprocessor based instantiation. There are still some bugs to work out in testing