Fixing MinMaxStat test to always write keys
Its still broken on table cleanup, but otherwise works. The problem was that the
RemoveRegionOnSplit coprocessor ran _after_ the other statistic coprocessors, so it
would return null from preCompactScannerOpen method, effectively dropping any
previous modifications from other coprocessors that ran before it. This made it
very finiky, but didn't gain anything because the default InternalScanner
passed in was null anyways.

TODO:
 - Fix table cleanup. Its still not removing the stats on table deletion,
but this is a minor issue and known from the original patch