Fixing index lock access + adding tests
Fixing issue around accidentally unlocking the index update lock when there are no updates.

Previously, when we attempt to make an index update, but there are no resulting updates
(i.e. delete of an empty row) we early out in doPre which skips taking the update lock
since we don't need to lock rolling of the WAL if there are no index updates to write.
However, we still release the lock in doPost, even if no index updates had been made.
Therefore, we end up getting an IllegalMonitorException since that lock is unheld.

Also, adding an extra test for the simple delete of a row with an existing index case.